Proper HTML escaping in view templates
We *need* to properly escape all view data output to avoid XSS attack
vectors.  The only exception are view data which are already supplied
as HTML.  We mark such values with the type alias `html`, and introduce
`View::raw()` to handle them.  Everything else should be passed to
`View::esc()` or be passed to one of the other View methods.

Apparently, SimplePie already yields escaped/sanitized HTML, so almost
all values of the FeedView templates are `html`.