Cater to possibly missing feed data
All SimplePie methods we're calling will return null, if the feed does
not contain the requested information.  Instead of silently converting
to empty strings, we pass the nulls all along, so the templates can
finally deal with it as desired; in the shipped template we don't show
the respective HTML fragments at all, but custom templates could use
`else` clauses to display something else, for example.